feat: approve miscellaneous expense/income

+++ src/transaction/transaction.service.ts
@@ -20,9 +20,13 @@ export class TransactionService extends BaseService<Trans> {
           account: true
         }
       },
-      store: true,
+      store: {
+        include: {
+          company: true
+        }
+      }
     }
-    super('trans', db, relations);
+    super('trans', db, relations, true);
   }
 
   async create(data: any) {
@@ -126,22 +130,22 @@ export class TransactionService extends BaseService<Trans> {
             ...updatedTrans,
           }
         });
-        console.log('udpatedTrans', trans);
         // Update the transaction details
-        // delete all transaction details before
-        const prevTransDetails = await this.db.trans_Details.deleteMany({
-          where: {
-            trans_id: id 
-          }
-        });
-        // then create new transaction details
-        const transDetails = await this.db.trans_Details.createMany({
-          data: accounts.map(account => ({
-            trans_id: id,
-            ...account
-          })),
-        });
-        console.log('updated TransDetails', transDetails);
+        if (accounts) {
+          // delete all transaction details before
+          const prevTransDetails = await this.db.trans_Details.deleteMany({
+            where: {
+              trans_id: id 
+            }
+          });
+          // then create new transaction details
+          const transDetails = await this.db.trans_Details.createMany({
+            data: accounts.map(account => ({
+              trans_id: id,
+              ...account
+            })),
+          });
+        }
       });
 
       return result;
@@ -153,19 +157,22 @@ export class TransactionService extends BaseService<Trans> {
 
 
   async delete(id: any) {
-    // delete trans_details with the id of the transaction
-    // then if success delete the transaction
     try {
+      // delete trans_details with the id of the transaction
       const result = await this.db.$transaction(async (prisma) => {
-        const transDetails = await this.db.trans_Details.deleteMany({
+        const transDetails = await this.db.trans_Details.updateMany({
           where: {
             trans_id: id
+          },
+          data: {
+            deleted_at: new Date()
           }
         });
         // delete the transaction
-        const trans = await this.db.trans.delete({
-          where: {
-            id: id
+        const trans = await this.db.trans.update({
+          where: { id: id },
+          data: {
+            deleted_at: new Date(),
           }
         })
       });
@@ -174,6 +181,74 @@ export class TransactionService extends BaseService<Trans> {
       console.error('Error deleting transaction:', error);
       throw new Error('Transaction deletion failed');
     }
+  }
+
+  async updateStatus(id: any, status: number) {
+    const trans = await this.findOne(id);
+
+    try {
+      // Status == 1/approved -> insert report journals
+      if (status == 1) {
+        trans.trans_details.forEach(async (row) => {
+          const data = {
+            trans_id: trans.id,
+            code: trans.code, 
+            company_id: trans.store.company.id,
+            company_name: trans.store.company.name,
+            store_id: trans.store_id,
+            store_name: trans.store.name,
+            trans_date: trans.trans_date,
+            trans_type_id: trans.trans_type_id,
+            trans_type_code: trans.trans_type.code,
+            trans_type_name: trans.trans_type.name,
+            description: trans.description,
+            account_id: row.account_id,
+            account_name: row.account.name,
+            amount: row.amount,
+            detail_description: row.description,
+            cash_bank: row.kas,
+          };
+          const result = await this.db.report_Journals.create({
+            data: {
+              trans_id: trans.id,
+              code: trans.code, 
+              company_id: trans.store.company.id,
+              company_name: trans.store.company.name,
+              store_id: trans.store_id,
+              store_name: trans.store.name,
+              trans_date: trans.trans_date,
+              trans_type_id: trans.trans_type_id,
+              trans_type_code: trans.trans_type.code,
+              trans_type_name: trans.trans_type.name,
+              description: trans.description,
+              account_id: row.account_id,
+              account_name: row.account.name,
+              amount: row.amount,
+              detail_description: row.description,
+              cash_bank: row.kas,
+            }
+          });
+  
+        });
+      }
+      // Status == 0/ rejected -> deleted from  report journals
+      else  if (status == 0) {
+        // Delete from report journals
+        const result = await this.db.report_Journals.deleteMany({
+          where: {
+            trans_id: trans.id,
+          }
+        });
+      }
+
+    } catch (error) {
+      return ResponseDto.error('Error updating transaction status', null);
+    }
+
+    // Update trans
+    const updatedTrans = await this.update(id, { approve: status });
+  
+    return updatedTrans;
   }
 
   async countThisMonthTransaction(store_id: string, year: number, month: number, trans_type_id: number) {
@@ -231,6 +306,7 @@ export class TransactionService extends BaseService<Trans> {
           a.name as account_name, 
           ABS(t.total) as total, 
           t.description as description, 
+          t.approve as approve,
           t.created_by as created_by,
           td.kas
         FROM "Trans" t
@@ -239,8 +315,8 @@ export class TransactionService extends BaseService<Trans> {
         WHERE td.kas = true
       `;
   
-      if (filters.account_name && filters.account_name !== '' && filters.account_name.length > 0) {
-        query = Prisma.sql`${query} AND td.account_id = ${filters.account_name[0]}::uuid`;
+      if (filters.account_id && filters.account_od !== '') {
+        query = Prisma.sql`${query} AND td.account_id = ${filters.account_id}::uuid`;
       }
       if (filters.trans_type_id && filters.trans_type_id !== '') {
         query = Prisma.sql`${query} AND t.trans_type_id = ${filters.trans_type_id}`;
@@ -253,6 +329,8 @@ export class TransactionService extends BaseService<Trans> {
         // Directly pass the formatted date as a string
         query = Prisma.sql`${query} AND t.trans_date <= ${filters.end_date}::date`;
       }
+
+      query = Prisma.sql`${query} AND t.deleted_at IS NULL`;
   
       const result = await this.db.$queryRaw(query);
   
@@ -264,67 +342,229 @@ export class TransactionService extends BaseService<Trans> {
   }
 
   async createSales(data: any) {
-    // const transType = await this.transTypeService.findOne({ where: { code: 'SAL' } });
-    // const transTypeId = transType.id;
-    // const hasTax = true // TODOELLA: Check if the store has tax
-    // const store = await this.db.stores.findUnique({ where: { id: data.store_id } });
-    // var transDetailsFormated = [];
+    // Kas / Piutang Usaha     (D)  9.990  
+    // Diskon Penjualan        (D)  1.000  
+    // Pendapatan                         (K) 10.000  (gk termasuk diskon)
+    // Utang Pajak (PPN)                  (K)    990  
+    const transType = await this.transTypeService.findOne({ where: { code: 'SAL' } });
+    const transTypeId = transType.id;
+    const hasTax = true // TODOELLA: Check if the store has tax
+    const store = await this.db.stores.findUnique({ where: { id: data.store_id } });
+    var diskonTotal = 0;
+    var salesEmasTotal = 0;
+    data.store = store;
+    var transDetailsFormated = [];
 
-    // // REFORMAT TRANSACTION DETAIL
-    // // Tax 
-    // if (hasTax) {
-    //   const taxAccount = await this.db.trans_Account_Settings.findUnique({where: {
-    //     company_id: data.auth.company_id,
-    //     action: 'tax'
-    //   }})
-    //   transDetailsFormated.push({
-    //     account_id: taxAccount.account_id,
-    //     amount:  Math.abs(data.tax_price) * -1,
-    //     description: 'Pajak Penjualan',
-    //     kas: false
-    //   })
-    // }
-    // // Details
-    // data.transaction_details.forEach(det => {
-    //   // Sales Operation
-    //   if ("operation_id" in det) {
-    //     transDetailsFormated.push({
-    //       account_id: det.account_id,
-    //       amount: Math.abs(det.amount),
-    //       description: det.description,
-    //       kas: true
-    //     })
-    //   } 
-    //   // Sales Emas
-    //   else {
+    // REFORMAT TRANSACTION DETAIL
+    // JOURNAL ENTRY (KREDIT)
+    // Tax yg dibayar customer
+    if (hasTax) {
+      const taxAccount = await this.db.trans_Account_Settings.findUnique({
+        where: {
+          store_id_company_id_action: { 
+            store_id: null, 
+            company_id: data.auth.company_id, 
+            action: 'tax' 
+          }
+        }
+      });
+      
+      transDetailsFormated.push({
+        account_id: taxAccount.account_id,
+        amount:  Math.abs(data.tax_price) * -1,
+        description: 'Pajak Penjualan',
+        kas: false
+      })
+    }
+    // Details / Journal Entry operation
+    data.transaction_details.forEach(det => {
+      // Sales Operation
+      if ("operation_id" in det) {
+        transDetailsFormated.push({
+          account_id: det.account_id,
+          amount: Math.abs(det.total_price) * -1,
+          description: det.name,
+          kas: false,
+        })
+      } 
+      // Sales Emas
+      else {
+        salesEmasTotal += (Math.abs(det.total_price) + Math.abs(det.discount)) * -1;
+        diskonTotal += Math.abs(det.discount);
+      }
+    })
+    // Journal entry sales emas
+    const goldSalesAccount = await this.getGoldSalesAccount(data);
+    transDetailsFormated.push({
+      account_id: goldSalesAccount.account_id,
+      amount: salesEmasTotal,
+      description: 'Penjualan Emas',
+      kas: false,
+    })
 
-    //   }
-    // })
+    // JOURNAL ENTRY (DEBIT)
+    // Diskon Penjualan
+    const discountAccount = await this.getDiscountAccount(data);
+    if (diskonTotal > 0) {
+      transDetailsFormated.push({
+        account_id: discountAccount.account_id,
+        amount: diskonTotal,
+        description: 'Diskon penjualan',
+        kas:true        
+      })
+    }
+    // Account Debit
+    // payment_method   Int // 1: Cash, 2: Bank Transfer, 3: Credit Card, 4: Debit Card
+    
 
-    // // CREATE TRANSACTION
-    // try {
-    //   const trans = await this.db.trans.create({
-    //     data: {
-    //       code: data.code,
-    //       store: { connect: { id: data.store_id } },
-    //       trans_type: { connect: { id: transTypeId } },
-    //       total : data.total_price,
-    //       description: store.name + ' Sales',
-    //       trans_date: new Date(),
-    //       weight_total: data.weight_total,
-    //       sub_total_price: data.sub_total_price,
-    //       tax_price: data.tax_price,
-    //       created_by: data.employee_id,
-    //       updated_by: null,
+    // CREATE TRANSACTION
+    try {
+      const trans = await this.db.trans.create({
+        data: {
+          code: data.code,
+          store: { connect: { id: data.store_id } },
+          trans_type: { connect: { id: transTypeId } },
+          total : data.total_price,
+          description: store.name + ' Sales',
+          trans_date: new Date(),
+          weight_total: data.weight_total,
+          sub_total_price: data.sub_total_price,
+          tax_price: data.tax_price,
+          created_by: data.employee_id,
+          updated_by: null,
           
-    //       trans_details: {
-    //         create: {}
-    //       }
-    //     }
-    //   })
+          // trans_details: {
+          //   create: {}
+          // }
+        }
+      })
 
-    // }catch (error) {
-    //   return ResponseDto.error('Error creating sales transaction',null);
-    // }
+    }catch (error) {
+      return ResponseDto.error('Error creating sales transaction',null);
+    }
+
+    // INPUT TO JOURNAL
+  }
+
+
+  async getGoldSalesAccount(data: any) {
+    var goldSalesAccount = await this.db.trans_Account_Settings.findUnique({
+      where: {
+        store_id_company_id_action: { 
+          store_id: data.store_id, 
+          company_id: data.auth.company_id, 
+          action: 'goldSales' 
+        }
+      }
+    });
+    // create an account and set is as gold sales what if foldS
+    if (!goldSalesAccount) {
+      // Generate Account Code
+      var lastCodeForPendapatan = await this.db.accounts.findMany({
+        where: {
+          company_id: data.store.company_id,
+          account_type_id: 5
+        },
+        orderBy: {
+          code: 'desc'
+        }
+      });
+      var codePendapatan = 40000;
+      if (lastCodeForPendapatan.length > 0) {
+        var lastCode = lastCodeForPendapatan[0].code;
+        codePendapatan = lastCode + 1;
+      }
+      // Create new account
+      const newAccount = await this.db.accounts.create({
+        data: {
+          code: codePendapatan,
+          name: 'Penjualan Emas ' + data.store.name,
+          account_type: { connect: { id: 5 } },
+          description: 'Default Akun Penjualan ' + data.store.name,
+          store: { connect: { id: data.store_id } },
+          company: { connect: { id: data.auth.company_id } },
+          deactive: false,
+          // created_by: data.employee_id,
+        }
+      })
+
+      // Assign to trans_account_settings 
+      goldSalesAccount = await this.db.trans_Account_Settings.create({
+        data: {
+          store: {
+            connect: { id: data.store_id }
+          },
+          company: {
+            connect: {id: data.auth.company_id}
+          },
+          account: {
+            connect: {id: newAccount.id}
+          },
+          description: 'Default Akun Penjualan ' + data.store.name,
+          action: 'goldSales'
+        }
+      });
+    }
+
+    return goldSalesAccount;
+  }
+
+  async getDiscountAccount(data: any) {
+    var discountAccount = await this.db.trans_Account_Settings.findUnique({
+      where: {
+        store_id_company_id_action: { 
+          store_id: data.store_id, 
+          company_id: data.auth.company_id, 
+          action: 'discountSales' 
+        }
+      }
+    });
+    if (!discountAccount)  {
+      // Create new pendapatan account
+      var lastCodeForPendapatan = await this.db.accounts.findMany({
+          where: {
+            company_id: data.store.company_id,
+            account_type_id: 5
+          },
+          orderBy: {
+            code: 'desc'
+          }
+      });
+      var codePendapatan = 40000;
+      if (lastCodeForPendapatan.length > 0) {
+        var lastCode = lastCodeForPendapatan[0].code;
+        codePendapatan = lastCode + 1;
+      }
+      const newAccount = await this.db.accounts.create({
+        data: {
+          code: codePendapatan,
+          name: 'Penjualan Emas ' + data.store.name,
+          account_type: { connect: { id: 5 } },
+          description: 'Default Akun Penjualan ' + data.store.name,
+          store: { connect: { id: data.store_id } },
+          company: { connect: { id: data.auth.company_id } },
+          deactive: false,
+          // created_by: data.employee_id
+        }
+      });
+      // Assign to trans_account_settings 
+      discountAccount = await this.db.trans_Account_Settings.create({
+        data: {
+          store: {
+            connect: { id: data.store_id }
+          },
+          company: {
+            connect: {id: data.auth.company_id}
+          },
+          account: {
+            connect: {id: newAccount.id}
+          },
+          description: 'Default Akun Diskon Penjualan ' + data.store.name,
+          action: 'discountSales'
+        }
+      });
+    }
+
+    return discountAccount;
   }
 }
